pyproject.toml
[tool.poetry]
name = "Fastapi_Auth"
version = "0.1.0"
description = ""
authors = ["muhammadddev <muhammadddev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.66.0"
Hypercorn = {extras = ["uvloop"], version = "^0.11.2"}
uvloop = "^0.15.3"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
SQLAlchemy = "^1.4.17"
arrow = "^1.1.0"
environ-config = "^21.2.0"
email-validator = "^1.1.2"
psycopg2 = "^2.9.1"
python-multipart = "^0.0.5"
alembic = "^1.6.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = {extras = ["d"], version = "^21.7b0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
